deleted unused UserDto and Mapper from event package, refactored imports
