+++ src/main/java/school/faang/user_service/service/event/EventParticipationService.java
@@ -2,8 +2,8 @@ package school.faang.user_service.service.event;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import school.faang.user_service.dto.event.UserDto;
-import school.faang.user_service.mapper.event.UserMapper;
+import school.faang.user_service.dto.user.UserDto;
+import school.faang.user_service.mapper.user.UserMapper;
 import school.faang.user_service.repository.event.EventParticipationRepository;
 
 import java.util.List;
